Move pcap test helpers to handlers_pcap_test.go
